+++ testbench_xilinx/TbUart.vhd
@@ -97,7 +97,7 @@ begin
 
   ------------------------------------------------------------
   -- create Clock 
-  Osvvm.TbUtilPkg.CreateClock ( 
+  Osvvm.ClockResetPkg.CreateClock ( 
   ------------------------------------------------------------
     Clk        => Clk, 
     Period     => tperiod_Clk 
@@ -105,7 +105,7 @@ begin
   
   ------------------------------------------------------------
   -- create nReset 
-  Osvvm.TbUtilPkg.CreateReset ( 
+  Osvvm.ClockResetPkg.CreateReset ( 
   ------------------------------------------------------------
     Reset       => nReset,
     ResetActive => '0',